+++ Editor/editor-tools/UnityFile.cs
@@ -7,29 +7,53 @@ using Cobilas.Unity.Utility;
 namespace Cobilas.Unity.Editor.Utility {
     public static class UnityFile {
 
-        [MenuItem("Assets/Create/Cobilas file/Text file")]
+        [MenuItem("Assets/Create/Unity file/Text file")]
         public static void InitTextFile()
             => CreateFile("TEXTFile.txt");
 
-        [MenuItem("Assets/Create/Cobilas file/Global custom defines/Target .NET 3.5 ")]
+        [MenuItem("Assets/Create/Unity file/Global custom defines/Target .NET 3.5 ")]
         public static void InitGlobalCustomDefinesNET3_5()
             => CreateFile("mcs.rsp");
 
-        [MenuItem("Assets/Create/Cobilas file/Global custom defines/Target .NET 4.x ")]
+        [MenuItem("Assets/Create/Unity file/Global custom defines/Target .NET 4.x ")]
         public static void InitGlobalCustomDefinesNET4_X()
             => CreateFile("csc.rsp");
 
-        [MenuItem("Assets/Create/Cobilas file/XML file")]
+        [MenuItem("Assets/Create/Unity file/XML file")]
         public static void InitXMLFile()
             => CreateFile("XNLFile.xml",
                 "<?xml version=\"1.0\" encoding=\"utf - 8\"?>\n"
                 );
 
-        [MenuItem("Assets/Create/Cobilas file/EmpytCS file")]
+        [MenuItem("Assets/Create/Unity file/EmpytCS file")]
         public static void InitEmpCsharpFile()
             => CreateFile("EmpytCS.cs",
                 ""
                 );
+
+        [MenuItem("Assets/Create/Unity file/CSharp Editor file")]
+        public static void InitEditorCSharpFile() {
+            StringBuilder build = new StringBuilder();
+            build.AppendLine("using UnityEngine;");
+            build.AppendLine("using UnityEditor;\n");
+            build.AppendLine("public class CSharpEditor : Editor {\n\n}");
+            CreateFile("CSharpEditor.cs", build.ToString());
+        }
+
+        [MenuItem("Assets/Create/Unity file/CSharp EditorWindow file")]
+        public static void InitEditorWinCSharpFile() {
+            StringBuilder build = new StringBuilder();
+            build.AppendLine("using UnityEngine;");
+            build.AppendLine("using UnityEditor;\n");
+            build.AppendLine("public class CSharpEditorWindow : EditorWindow {\n");
+            build.AppendLine("\t[MenuItem(\"Window/CSharpEditorWindow\")]");
+            build.AppendLine("\tprivate static void Init() {");
+            build.AppendLine("\t\tCSharpEditorWindow temp = GetWindow<CSharpEditorWindow>();");
+            build.AppendLine("\t\ttemp.titleContent = new GUIContent(\"CSharp Editor Window\");");
+            build.AppendLine("\t\ttemp.Show();");
+            build.AppendLine("\t}\n}");
+            CreateFile("CSharpEditorWindow.cs", build.ToString());
+        }
 
         private static void CreateFile(string newFile)
             => CreateFile(newFile, "");